\name{availableModels}
\alias{availableModels}
\title{Available models}
\description{
  Available built-in time series models
}
\usage{
availableModels()
}
\details{
  Return the list of built-in available \sQuote{nlar} time series models
}
\value{
A character vector containing built-in time series models. For help on a specific model, type:
\code{help(modelName)}.
}
\author{ Antonio, Fabio Di Narzo }
\examples{
availableModels()
}
\keyword{ ts }
