\name{barry}
\alias{barry}
\title{Time series of PPI used as example in Bierens and Martins (2010)}
\description{
  This data set contains the series used by Bierens and Martins 
  for testing for PPI between Canada and US.
}
\usage{data(barry)}
\format{
  A data frame with 324 monthly observations,
  ranging from 1973:M1 until 1999:M12.

    \tabular{rl}{
    \code{dolcan} \tab Exchange rate US/Can dollar. \cr
    \code{cpiUSA} \tab US Consumer Price Index. \cr
    \code{cpiCAN} \tab Canada Consumer Price Index. \cr
    }

}
\source{
  Bierens, H. and Martins, L. (2010), Time Varying Cointegration,
}
\author{Matthieu Stigler}
\keyword{datasets}
