\name{IIPUs} 
\alias{IIPUs}
\docType{data}
\title{US monthly industrial production from Hansen (1999)}
\description{
  This data, used as example in Hansen (1999), contains the US monthly industrial production.
}
\usage{data(IIPUs)}
\format{
  A monthly time series of class ts starting in january 1960 and ending in september 1998
}
\source{Hansen (1999) Testing for linearity, Journal of Economic Surveys, Volume 13, Number 5, December 1999 , pp. 551-576(26)
avalaible at: \url{http://www.ssc.wisc.edu/~bhansen/papers/cv.htm}
}\keyword{datasets}
\concept{Testing for linearity}

