% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nlar-methods.R
\name{MAPE}
\alias{MAPE}
\alias{MAPE.default}
\title{Mean Absolute Percent Error}
\usage{
MAPE(object, ...)

\method{MAPE}{default}(object, ...)
}
\arguments{
\item{object}{object of class \code{nlar.fit}}

\item{\dots}{additional arguments to \code{MAPE}}
}
\value{
Computed Mean Absolute Percent Error for the fitted model.
}
\description{
Generic function to compute the Mean Absolute Percent Error of a fitted
model.
}
\author{
Antonio, Fabio Di Narzo
}
\keyword{ts}

