\name{mse}
\alias{mse}
\alias{mse.default}
\title{ Mean Square Error }
\description{
  Generic function to compute the Mean Squared Error of a fitted model.
}
\usage{
mse(object, ...)

## Default S3 method:
mse(object, ...)
}
\arguments{
  \item{object}{ object of class \code{nlar.fit} }
  \item{\dots}{ additional arguments to \code{mse} }
}
\value{
  Computed MSE for the fitted model.
}
\author{Antonio, Fabio Di Narzo}
\keyword{ ts }
