\name{nlar.struct}
\alias{nlar.struct}
\title{NLAR common structure}
\description{
NLAR common structure
}
\usage{
nlar.struct(x, m, d=1, steps=d, series)
}
\value{
  An object of class \code{nlar.struct}.
}
\arguments{
\item{x}{time series}
\item{m,d,steps}{embedding dimension, time delay and forecasting steps}
\item{series}{(optional) time series name}
}
\author{ Antonio, Fabio Di Narzo }
\keyword{ ts }
\keyword{ internal }
