\name{predict}

\alias{predict}

\alias{predict.nlar}

\encoding{latin1}

\title{Predict method for objects of class nlar}

\description{
  Forecating a non-linear model object of general class \sQuote{\code{nlar}}, including \sQuote{\code{setar}} and \sQuote{\code{star}}.
}

\usage{
\method{predict}{nlar}(object, newdata, n.ahead=1, type=c("naive", "MC", "bootstrap"), nboot=100, ci=0.95, ...)
}
\arguments{
  \item{object}{An object of class \sQuote{\code{nlar}}; generated by
    \command{setar()} or \command{lstar()}.} 
  \item{newdata}{Optional. A new data frame to predict from.} 
  \item{n.ahead}{An integer specifying the number of forecast steps.}
  \item{type}{Type of forecasting method used. See details.}
  \item{nboot}{The number of replications for type MC or bootstrap.} 
  \item{ci}{The forecast confidence interval (available only with types MC and bootstrap). }
  \item{\dots}{Currently not used.}
}
\details{
  The forecasts are obtained recursively from the estimated model. 
Given that the models are non-linear, ignoring the residuals in the 2- 
and more steps ahead forecasts leads to biased forecasts (so-called naive). 
Different resampling methods, averaging \code{n.boot} times over future residuals, are available:

\describe{
  \item{naive}{No residuals}
  \item{MC}{Monte-Carlo method, where residuals are taken from a normal distribution, with sd. equal to the residuals sd. }
  \item{bootstrap}{Residuals are resampled from the empirical residuals from the model.}
}
}

\value{
  A \sQuote{\code{ts}} object, or, in the case of MC/bootstrap, a list containing the prediction (pred) and the forecast standard errors (\code{se}).
}

\author{Matthieu Stigler}


\seealso{
  \code{\link{setar}}, \code{\link{lstar}}
} 

\examples{

x.train <- window(log10(lynx), end = 1924)
x.test <- window(log10(lynx), start = 1925)


mod.set <- setar(x.train, m=2, thDelay=0)
pred_setar_naive <- predict(mod.set, n.ahead=10)
pred_setar_boot <- predict(mod.set, n.ahead=10, type="bootstrap", n.boot=200)

## Plot to compare results:
plot(x.test)
lines(pred_setar_naive, lty=2, col=2)
lines(pred_setar_boot$pred, lty=3, col=3)

}

\keyword{regression}
\concept{VAR}
\concept{VECM}
\concept{Vector autoregressive}
\concept{Forecasts of VAR}
\concept{Prediction of VAR}
