% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_datasets.R
\docType{data}
\name{setarTest_IIPUs_results}
\alias{setarTest_IIPUs_results}
\title{Results from the setarTest, applied on Hansen (1999) data}
\format{
A list containing output from two tests
}
\usage{
setarTest_IIPUs_results
}
\description{
Saved objects
}
\examples{
library(tsDyn)
data(IIPUs)
B <-  1000
\dontrun{
  test_1 <- setarTest(IIPUs, m=16, thDelay=5, nboot=B)
  test_2 <- setarTest(IIPUs, m=16, thDelay=5, nboot=B, test = "2vs3")
}
}
\seealso{
Example in \code{\link{IIPUs}}
}
\keyword{datasets}
