% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sigmoid}
\alias{d2sigmoid}
\alias{dsigmoid}
\alias{sigmoid}
\title{sigmoid functions}
\usage{
sigmoid(x)

dsigmoid(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Some sigmoid functions. See R sources for their definition
}
\author{
J. L. Aznarte
}
\keyword{ts}

