\name{addRegime}
\alias{addRegime}
\title{addRegime test}
\description{
addRegime test
}
\usage{
addRegime(object)
}
\arguments{
\item{object}{ fitted model object with at least 2 regimes }
}
\value{
A list containing the p-value of the F statistic and a boolean,
true if there is some remaining nonlinearity and false otherwise.
}
\seealso{\code{\link{star}}}
\author{ J. L. Aznarte }
\examples{
##TODO
}
\keyword{ ts }
\references{
TODO
}
