\name{addRegime}
\alias{addRegime}
\title{addRegime test}
\usage{
  addRegime(object, ...)
}
\arguments{
  \item{object}{fitted model object with at least 2
  regimes}

  \item{...}{arguments to and from other methods}
}
\value{
  A list containing the p-value of the F statistic and a
  boolean, true if there is some remaining nonlinearity and
  false otherwise.
}
\description{
  addRegime test
}
\examples{
##TODO
}
\author{
  J. L. Aznarte
}
\references{
  TODO
}
\seealso{
  \code{\link{star}}
}
\keyword{ts}

