% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotriples.rgl.R
\name{autotriples.rgl}
\alias{autotriples.rgl}
\title{Interactive trivariate time series plots}
\usage{
autotriples.rgl(x, lags = 1:2, type = c("lines", "points"))
}
\arguments{
\item{x}{time series}

\item{lags}{vector of regressors lags}

\item{type}{type of plot: contour levels, perspective plots, image}
}
\value{
None. A plot is produced on the current \code{rgl} device.
}
\description{
Interactive trivariate time series plots
}
\details{
This function displays interactive trivariate time series plots
\code{x[t-lags[1]], x[t-lags[2]]} against \code{x[t]} using the interactive
\code{\link[rgl]{rgl}} device.
}
\examples{

if(interactive())
autotriples.rgl(log(lynx))

}
\seealso{
\code{\link{autotriples}} for 3d visualization via
\pkg{scatterplot3d} package and for kernel post-processing of the cloud for
nonparametric autoregression functions estimates.
}
\author{
Wrapper to 'sm' and GUI by Antonio, Fabio Di Narzo
}
\keyword{ts}
