% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{availableModels}
\alias{availableModels}
\title{Available models}
\usage{
availableModels()
}
\value{
A character vector containing built-in time series models. For help
on a specific model, type: \code{help(modelName)}.
}
\description{
Available built-in time series models
}
\details{
Return the list of built-in available \sQuote{nlar} time series models
}
\examples{
availableModels()
}
\author{
Antonio, Fabio Di Narzo
}
\keyword{ts}

