\name{computeGradient}
\alias{computeGradient}
\title{computeGradient}
\description{
Computes the gradient under the null hypothesis
}
\usage{
computeGradient(object, ...)
}
\arguments{
\item{object}{ fitted star model }
\item{...}{currently unused}
}
\value{
computed gradient
}
\seealso{\code{\link{addRegime}}}
\author{ J. L. Aznarte }
\examples{
##TODO
}
\keyword{ ts }
\references{
TODO
}
