% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/star.R
\name{computeGradient}
\alias{computeGradient}
\title{computeGradient}
\usage{
computeGradient(object, ...)
}
\arguments{
\item{object}{fitted star model}

\item{...}{currently unused}
}
\value{
computed gradient
}
\description{
Computes the gradient under the null hypothesis
}
\examples{
##TODO
}
\author{
J. L. Aznarte
}
\references{
TODO
}
\seealso{
\code{\link{addRegime}}
}
\keyword{internal}
\keyword{ts}

