\name{UsUnemp} 
\alias{UsUnemp}
\docType{data}
\title{US unemployment series used in Caner and Hansen (2001)}
\description{
  This data, used as example in Caner and Hansen (2001), contains the montly US adult male unemployment from 1956 to 1999.
}
\usage{data(UsUnemp)}
\format{
  A monthly time series of class ts starting in January 1956 and ending in August 1999.}
\source{
Caner and Hansen,Threshold autoregression with a unit root Econometrica, 2001, 69, 1555-1596
avalaible at: \url{http://www.ssc.wisc.edu/~bhansen/papers/cv.htm}
}
\keyword{datasets}
\concept{Testing for linearity}

