% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIRF.R
\name{GIRF}
\alias{GIRF}
\alias{GIRF.setar}
\alias{GIRF.linear}
\alias{GIRF.nlVar}
\title{Generalized Impulse response Function (GIRF)}
\usage{
GIRF(object, n.ahead, seed = NULL, ...)

\method{GIRF}{setar}(object, n.ahead = 10, seed = NULL, n.hist = 20,
  n.shock = 20, R = 10, hist_li = NULL, shock_li = NULL, ...)

\method{GIRF}{linear}(object, n.ahead = 10, seed = NULL, n.hist = 20,
  n.shock = 20, R = 10, hist_li = NULL, shock_li = NULL, ...)

\method{GIRF}{nlVar}(object, n.ahead = 10, seed = NULL, n.hist = 20,
  n.shock = 20, R = 10, hist_li = NULL, shock_li = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{linear}} or \code{\link{setar}}}

\item{n.ahead}{The numher of steps ahead to compute}

\item{seed}{optional, the seed for the random numbers}

\item{n.hist}{The number of past histories ot consider}

\item{n.shock}{The number of actual shocks ot consider}

\item{R}{the number of draws to use for the n.ahead innovations}

\item{hist_li}{optional, a list of histories (each of same length as lags in the model)}

\item{shock_li}{optional, a list of innovations}

\item{\ldots}{Further argumetns passed to specific methods.}
}
\description{
Generates a GIRF for multiple innovations and histories
}
