% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isLinear.R
\name{isLinear}
\alias{isLinear}
\title{isLinear}
\usage{
isLinear(object, ...)
}
\arguments{
\item{object}{fitted time series model}

\item{...}{arguments to and from other methods}
}
\description{
Generic NLAR linearity test
}
\author{
A. F. Di Narzo
}
\keyword{ts}

