% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lineVar.R
\name{lineVar}
\alias{lineVar}
\title{Multivariate linear models: VAR and VECM}
\usage{
lineVar(data, lag, r = 1, include = c("const", "trend", "none", "both"),
  model = c("VAR", "VECM"), I = c("level", "diff", "ADF"), beta = NULL,
  estim = c("2OLS", "ML"), LRinclude = c("none", "const", "trend", "both"),
  exogen = NULL)
}
\arguments{
\item{data}{multivariate time series (first row being first=oldest value)}

\item{lag}{Number of lags to include in each regime}

\item{r}{Number of cointegrating relationships}

\item{include}{Type of deterministic regressors to include}

\item{model}{Model to estimate. Either a VAR or a VECM}

\item{I}{For VAR only: whether in the VAR the variables are to be taken in
levels (original series) or in difference, or similarly to the univariate ADF
case.}

\item{beta}{for VECM only: imposed cointegrating value. If null, will be estimated}

\item{estim}{Type of estimator for the VECM: '2OLS' for the two-step approach
or 'ML' for Johansen MLE}

\item{LRinclude}{Possibility to include in the long-run relationship and the
ECT trend, constant... Can also be a matrix with exogeneous regressors}

\item{exogen}{Inclusion of exogenous variables (first row being first=oldest
value). Is either of same size than data (then automatically cut) or than
end-sample.}
}
\value{
Fitted model data
}
\description{
Estimate either a VAR or a VECM.
}
\details{
This function provides basic functionalities for VAR and VECM models. More
comprehensive functions are in package \pkg{vars}. A few differences appear
in the VECM estimation: \describe{
\item{Engle-Granger estimator}{The Engle-Granger estimator is available}
\item{Presentation}{Results are printed in a different ways, using a matrix form}
\item{lateX export}{The matrix of coefficients can be exported to latex,
       with or without standard-values and significance stars}
}

Two estimators are available: the Engle-Granger two step
approach (\code{2OLS}) or the Johansen (\code{ML}). For the 2OLS,
deterministic regressors (or external variables if \code{LRinclude} is of
class numeric) can be added for the estimation of the cointegrating value and
for the ECT. This is only working when the beta value is not pre-specified.

The argument \code{beta} is only for \code{\link{VECM}}, look at the specific help page for more details.
}
\examples{
data(zeroyld)
data<-zeroyld

#Fit a VAR
VAR<-lineVar(data, lag=1)
VAR
summary(VAR)

#compare results with package vars:
if(require(vars)) {
	a<-VAR(data, p=1)
	coef_vars <- t(sapply(coef(a), function(x) x[c(3,1,2),1]))
	all.equal(coef(VAR),coef_vars, check.attributes=FALSE)
}

###VECM
VECM.EG<-lineVar(data, lag=2, model="VECM")
VECM.EG
summary(VECM.EG)

VECM.ML<-lineVar(data, lag=2, model="VECM", estim="ML")
VECM.ML
summary(VECM.ML)


###Check Johansen MLE
myVECM<-lineVar(data, lag=1, include="const", model="VECM", estim="ML")
summary(myVECM, digits=7)
#comparing with vars package
if(require(vars)){
	a<-ca.jo(data, spec="trans")
	summary(a)
	#same answer also!
}

##export to Latex
toLatex(VECM.EG)
toLatex(summary(VECM.EG))
options("show.signif.stars"=FALSE)
toLatex(summary(VECM.EG), parenthese="Pvalue")
options("show.signif.stars"=TRUE)
}
\author{
Matthieu Stigler
}
\seealso{
\code{\link{VECM}} which is just a wrapper for
\code{lineVar(...,model="VECM")}

\code{\link{TVAR}} and \code{\link{TVECM}} for the corresponding threshold
models. \code{\link{linear}} for the univariate AR model.
}
\keyword{ts}

