\name{m.unrate}
\docType{data}
\alias{m.unrate}
\title{Monthly US unemployment }
\description{
  Monthly US civilian unemployment rate from 1948 through 2004.  
  Used in the book financial time series for Tsay (2005, chapter 4)
}
\usage{
data(m.unrate)
}
\format{
      zoo object 
}
\source{
  \samp{https://faculty.chicagobooth.edu/ruey-s-tsay/research/analysis-of-financial-time-series-2nd-edition}
}
\references{
  Ruey Tsay (2005) Analysis of Financial Time Series, 2nd ed. (Wiley,
  ch. 4)
}
\keyword{datasets}

