\name{MAPE}
\alias{MAPE}
\alias{MAPE.default}
\title{ Mean Absolute Percent Error }
\description{
  Generic function to compute the Mean Absolute Percent Error of a fitted model.
}
\usage{
MAPE(object, ...)

## Default S3 method:
MAPE(object, ...)
}
\arguments{
  \item{object}{ object of class \code{nlar.fit} }
  \item{\dots}{ additional arguments to \code{MAPE} }
}
\value{
  Computed MAPE for the fitted model.
}
\author{Antonio, Fabio Di Narzo}
\keyword{ ts }
