% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlar-methods.R
\name{mse}
\alias{mse}
\alias{mse.default}
\title{Mean Square Error}
\usage{
mse(object, ...)

\method{mse}{default}(object, ...)
}
\arguments{
\item{object}{object of class \code{nlar.fit}}

\item{\dots}{additional arguments to \code{mse}}
}
\value{
Computed MSE for the fitted model.
}
\description{
Generic function to compute the Mean Squared Error of a fitted model.
}
\author{
Antonio, Fabio Di Narzo
}
\keyword{ts}
