% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict.VAR.R, R/predict.VECM.R
\name{predict.VAR}
\alias{predict.VAR}
\alias{predict.VECM}
\title{Predict method for objects of class \sQuote{\code{VAR}} or \sQuote{\code{VECM}}}
\usage{
\method{predict}{VAR}(object, newdata, n.ahead = 5, newdataTrendStart,
  exoPred = NULL, ...)

\method{predict}{VECM}(object, newdata, n.ahead = 5, newdataTrendStart,
  exoPred = NULL, ...)
}
\arguments{
\item{object}{An object of class  \sQuote{\code{VAR}} or \sQuote{\code{VECM}}}

\item{newdata}{Optional. A new data frame to predict from.}

\item{n.ahead}{An integer specifying the number of forecast steps.}

\item{newdataTrendStart}{If \sQuote{\code{newdata}} is provided by the user,
and the estimated model includes a trend,
this argument specifies where the trend should start}

\item{exoPred}{vector/matrix of predictions for the exogeneous variable(s)}

\item{\dots}{Arguments passed to the unexported \sQuote{\code{VAR.gen}} function}
}
\value{
A matrix of predicted values.
}
\description{
Forecating a  \sQuote{\code{VAR}} / \sQuote{\code{VECM}}
}
\details{
The forecasts are obtained recursively from the estimated model.
}
\examples{
data(barry)
mod_vecm <- VECM(barry, lag=2)
predict(mod_vecm)

# note that when providing newdata, newdata has to be ordered chronologically,
# so that the first row/element is the earliest value:
all.equal(predict(mod_vecm), predict(mod_vecm, newdata=barry[c(322, 323, 324),]))
# unused so far, make sure;
}
\author{
Matthieu Stigler
}
\seealso{
\code{\link{lineVar}} and \code{\link{VECM}}.

A more sophisticated predict function, allowing to do sub-sample rolling
predictions: \code{\link{predict_rolling}}.
}
\keyword{regression}

