% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd_nlVar.R, R/irf.nlVar.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fevd}
\alias{irf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vars}{\code{\link[vars]{fevd}}, \code{\link[vars]{irf}}}
}}

