% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sigmoid}
\alias{sigmoid}
\alias{d2sigmoid}
\alias{dsigmoid}
\title{sigmoid functions}
\usage{
sigmoid(x)

dsigmoid(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Some sigmoid functions. See R sources for their definition
}
\author{
J. L. Aznarte
}
\keyword{ts}
