% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setar.R
\name{toLatex.setar}
\alias{toLatex.setar}
\title{Latex representation of fitted setar models}
\usage{
\method{toLatex}{setar}(object, digits = 3, label, ...)
}
\arguments{
\item{object}{fitted setar model (using \code{\link{nlar}})}

\item{digits}{options to be passed to \code{\link{format}} for formatting
numbers}

\item{label}{LaTeX label passed to the equation}

\item{\dots}{Not used}
}
\description{
Produce LaTeX output of the SETAR model.
}
\examples{
mod.setar <- setar(log10(lynx), m=2, thDelay=1, th=3.25)
toLatex(mod.setar)
}
\author{
Antonio, Fabio Di Narzo
}
\seealso{
\code{\link{setar}}, \code{\link{nlar-methods}}
}
\keyword{ts}

