\name{TVECM.sim}
\alias{TVECM.sim}
\title{Simulation and bootstrap of bivariate VECM/TVECM}
\description{
Estimate or bootstraps a multivariate Threshold VAR
}
\usage{
TVECM.sim(data,B,TVECMobject, nthresh=1, Thresh, beta, n=200, lag=1, type=c("simul","boot", "check"),  include = c("const", "trend","none", "both"), starting=NULL, sigma)
}


\value{
a list with the simulated/bootstraped data and the parameter matrix used. 
}
\arguments{
\item{data}{matrix of parameter to simulate}
\item{B}{Matrix of coefficients to simulate}
\item{TVECMobject}{Object computed by function TVECM() or linear VECM}
\item{nthresh}{number of threshold (see details)}
\item{Thresh}{The threshold value(s). Vector of length nthresh}
\item{beta}{The cointegrating value}
\item{n}{Number of observations to create when type="simul"}
\item{lag}{Number of lags to include in each regime}
\item{type}{Whether a bootstrap or simulation is to employ. See details}
\item{include}{Type of deterministic regressors to include. NOT WORKING PROPERLY CURRENTLY if not const}
\item{starting}{Starting values when a simulation with given parameter matrix is made}
\item{sigma}{variance of the errors}
}
\details{
This function offers the possibility to generate series following a TVECM from two approaches: bootstrap or simulation. 
When the parameter matrix is given, on can only simulate a VECM (nthresh=0) or TVECM (nthresh=1 or 2). One can have a specification with constant (default), trend, both or none (see arg include). Order in parameters is ECT/include/lags. The argument beta is the contegrating value on the right side of the long-run relationship, and hence the function use the vector (1, -beta). The starting values (of rows equal to the number of lags and cols equal to the number of variables) can be given. The user should take care for their choice, since it is not sure that the simulated values will cross the threshold even once. Notice that only one cointegrating value is allowed. User intrested in simulating a VECM with more cointegrating values should do use the VAR representation and use TVAR.sim. 
The second possibility is to bootstrap series. This is done on a object generated by TVECM (or VECM). A simple residual bootstrap is done, or one can simulate a serie with the same parameter matrix and with normal distributed residuals (with variance pre-specified), corresponding to Monte-carlo simulations. One can alternatively give only the series, and then the function will call internally TVECM()
}
\seealso{
TODO
}
\author{Matthieu Stigler}
\examples{
##Simulation of a TVAR with 1 threshold
B<-rbind(c(0.2, 0.11928245, 1.00880447, -0.009974585, 0.3, -0.089316, 0.95425564, 0.02592617),c( -0.1, 0.25283578, 0.09182279,  0.914763741, 0.35,-0.0530613, 0.02248586, 0.94309347))
sim<-TVECM.sim(B=B,beta=1, nthresh=1,n=500, type="simul",Thresh=5, starting=c(5.2, 5.5))$serie



##Bootstrap a TVAR with 1 threshold (two regimes)
data(zeroyld)
dat<-zeroyld
TVECMobject<-TVECM(dat, nthresh=1, lag=1, bn=20, ngridG=20, plot=FALSE)
TVECM.sim(TVECMobject=TVECMobject,type="boot")

##Check the bootstrap
all(TVECM.sim(TVECMobject=TVECM(dat, nthresh=1, lag=2, bn=20, ngridG=20, plot=FALSE, include="none"),type="check")$serie==dat)
}
\keyword{ ts }
\references{
TODO
}

